Commit 23.08.2018

+++ src/main/java/Model/Money.java
@@ -1,9 +1,11 @@
 package Model;
 
+import java.math.BigDecimal;
+
 public class Money {
 
     private String currency;
-    private double amount;
+    private BigDecimal amount;
 
     public String getCurrency() {
         return currency;
@@ -16,17 +18,17 @@ public class Money {
     public Money() {
     }
 
-    public Money(String currency, double amount) {
+    public Money(String currency, BigDecimal amount) {
 
         this.currency = currency;
         this.amount = amount;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
